# spring-user-service/src/main/java/dev/dmv04/userservice/dto/ErrorResponse.java
package dev.dmv04.userservice.dto;

import java.time.LocalDateTime;
import java.util.List;

public record ErrorResponse(
        LocalDateTime timestamp,
        int status,
        String error,
        String message,
        String path,
        List<ValidationError> details
) {}