# spring-user-service/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: User Service API
  description: API для управления пользователями с HATEOAS
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер

tags:
  - name: Users
    description: Операции с пользователями

paths:
  /api/users:
    get:
      tags:
        - Users
      summary: Получить всех пользователей
      description: Возвращает список всех зарегистрированных пользователей с HATEOAS ссылками
      responses:
        '200':
          description: Список пользователей с HATEOAS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCollection'

    post:
      tags:
        - Users
      summary: Создать нового пользователя
      description: Создает нового пользователя с указанными данными
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResource'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Получить пользователя по ID
      description: Возвращает данные пользователя по его уникальному идентификатору
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResource'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Обновить пользователя по ID
      description: Обновляет данные существующего пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Пользователь успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResource'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Удалить пользователя по ID
      description: Удаляет пользователя по его уникальному идентификатору
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      responses:
        '204':
          description: Пользователь успешно удалён
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: "Иван Иванов"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        age:
          type: integer
          minimum: 1
          example: 30
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2024-01-15T10:30:00"

    UserResource:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        age:
          type: integer
        createdAt:
          type: string
          format: date-time
        _links:
          $ref: '#/components/schemas/UserLinks'

    UserCollection:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            userDTOList:
              type: array
              items:
                $ref: '#/components/schemas/UserResource'
        _links:
          $ref: '#/components/schemas/CollectionLinks'

    UserLinks:
      type: object
      properties:
        self:
          $ref: '#/components/schemas/Link'
        update:
          $ref: '#/components/schemas/Link'
        delete:
          $ref: '#/components/schemas/Link'
        all-users:
          $ref: '#/components/schemas/Link'
        create-user:
          $ref: '#/components/schemas/Link'

    CollectionLinks:
      type: object
      properties:
        self:
          $ref: '#/components/schemas/Link'
        create-user:
          $ref: '#/components/schemas/Link'

    Link:
      type: object
      properties:
        href:
          type: string
          example: "http://localhost:8080/api/users/1"

    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - age
      properties:
        name:
          type: string
          minLength: 1
          example: "Иван Иванов"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        age:
          type: integer
          minimum: 1
          example: 25

    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          example: "Иван Иванов"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        age:
          type: integer
          minimum: 1
          example: 26

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
        path:
          type: string
          example: "/api/users"
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'

    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: "email"
        message:
          type: string
          example: "must be a valid email address"
        rejectedValue:
          type: string
          example: "invalid-email"